Don't attempt server submission if no server is specified

+++ daq.py
@@ -65,34 +65,38 @@ if __name__ == "__main__":
 
     newresults = [json.dump(s) for s in sensors]
 
-    try:
-        allresults = read_file(pendingfile)
-        allresults.extend(newresults)
+    if (config.has_option("daq", "server")):
+        try:
+            allresults = read_file(pendingfile)
+            allresults.extend(newresults)
 
-        if len(allresults) == 0:
+            if len(allresults) == 0:
+                sys.exit(0)
+
+            url = config.get("daq", "server")
+            if url[-1:] == "/":
+                url = url[:-1]
+            url = url + API_URL
+
+            result = urlopen(url, "\n".join(allresults))
+
+            if result.getcode() != 200:
+                raise Exception("Unexcepted HTTP response %d" % result.getcode())
+
+            processed = int(result.readline().strip())
+            if processed != len(allresults):
+                raise Exception("Didn't process all data packets, saw %d expected %d" % (processed, len(allresults)))
+
+            if config.has_option("daq", "backup") and (config.get("daq", "backup") == "true"):
+                filename = "backup-%s" % datetime.now().strftime("%Y%m%d%H%M%S")
+                while os.path.isfile(os.path.join(state, filename)):
+                    filename = filename + "-1"
+                os.rename(pendingfile, os.path.join(state, filename))
+            else:
+                os.remove(pendingfile)
+
             sys.exit(0)
+        except:
+            print_exc()
 
-        url = config.get("daq", "server")
-        if url[-1:] == "/":
-            url = url[:-1]
-        url = url + API_URL
-
-        result = urlopen(url, "\n".join(allresults))
-
-        if result.getcode() != 200:
-            raise Exception("Unexcepted HTTP response %d" % result.getcode())
-
-        processed = int(result.readline().strip())
-        if processed != len(allresults):
-            raise Exception("Didn't process all data packets, saw %d expected %d" % (processed, len(allresults)))
-
-        if config.has_option("daq", "backup") and (config.get("daq", "backup") == "true"):
-            filename = "backup-%s" % datetime.now().strftime("%Y%m%d%H%M%S")
-            while os.path.isfile(os.path.join(state, filename)):
-                filename = filename + "-1"
-            os.rename(pendingfile, os.path.join(state, filename))
-        else:
-            os.remove(pendingfile)
-    except:
-        append_file(pendingfile, newresults)
-        raise
+    append_file(pendingfile, newresults)
